# Stat benchmark: report cleanup time and optimise