@talex5 2024-02-14
Will investigate the deadlock separately; it's clearly not the fault of this PR.